feat(presentation): allow passing custom header component
