+++ packages/presentation/src/index.ts
@@ -9,11 +9,14 @@ export type {
   DocumentLocationsState,
   DocumentResolver,
   DocumentResolverContext,
+  HeaderOptions,
   NavigatorOptions,
   PresentationParams,
+  PresentationPerspective,
   PresentationPluginOptions,
   PresentationSearchParams,
   PresentationStateParams,
+  PresentationViewport,
   PreviewUrlOption,
   PreviewUrlResolver,
   PreviewUrlResolverOptions,
@@ -24,3 +27,18 @@ export {
   usePresentationNavigate,
 } from './usePresentationNavigate'
 export {usePresentationParams} from './usePresentationParams'
+export type {PreviewHeaderProps} from './preview/PreviewHeader'
+export type {PreviewProps} from './preview/Preview'
+export {
+  ACTION_IFRAME_LOADED,
+  ACTION_IFRAME_REFRESH,
+  ACTION_IFRAME_RELOAD,
+  ACTION_VISUAL_EDITING_OVERLAYS_TOGGLE,
+  type DispatchPresentationAction,
+  type IframeLoadedAction,
+  type IframeRefreshAction,
+  type IframeReloadAction,
+  type PresentationAction,
+  type PresentationState,
+  type VisualEditingOverlaysToggleAction,
+} from './reducers/presentationReducer'
